Refactor graph controls mode handling
